# ARROW-10318: [C++] Use pimpl idiom in CSV parser