--- a/cpp/src/arrow/csv/parser.h
+++ b/cpp/src/arrow/csv/parser.h
@@ -35,14 +35,89 @@ class MemoryPool;
 
 namespace csv {
 
-constexpr int32_t kMaxParserNumRows = 100000;
-
 /// Skip at most num_rows from the given input.  The input pointer is updated
 /// and the number of actually skipped rows is returns (may be less than
 /// requested if the input is too short).
 ARROW_EXPORT
 int32_t SkipRows(const uint8_t* data, uint32_t size, int32_t num_rows,
                  const uint8_t** out_data);
+
+class BlockParserImpl;
+
+namespace detail {
+
+struct ParsedValueDesc {
+  uint32_t offset : 31;
+  bool quoted : 1;
+};
+
+class ARROW_EXPORT DataBatch {
+ public:
+  explicit DataBatch(int32_t num_cols) : num_cols_(num_cols) {}
+
+  /// \brief Return the number of parsed rows
+  int32_t num_rows() const { return num_rows_; }
+  /// \brief Return the number of parsed columns
+  int32_t num_cols() const { return num_cols_; }
+  /// \brief Return the total size in bytes of parsed data
+  uint32_t num_bytes() const { return parsed_size_; }
+
+  template <typename Visitor>
+  Status VisitColumn(int32_t col_index, Visitor&& visit) const {
+    using detail::ParsedValueDesc;
+
+    for (size_t buf_index = 0; buf_index < values_buffers_.size(); ++buf_index) {
+      const auto& values_buffer = values_buffers_[buf_index];
+      const auto values = reinterpret_cast<const ParsedValueDesc*>(values_buffer->data());
+      const auto max_pos =
+          static_cast<int32_t>(values_buffer->size() / sizeof(ParsedValueDesc)) - 1;
+      for (int32_t pos = col_index; pos < max_pos; pos += num_cols_) {
+        auto start = values[pos].offset;
+        auto stop = values[pos + 1].offset;
+        auto quoted = values[pos + 1].quoted;
+        ARROW_RETURN_NOT_OK(visit(parsed_ + start, stop - start, quoted));
+      }
+    }
+    return Status::OK();
+  }
+
+  template <typename Visitor>
+  Status VisitLastRow(Visitor&& visit) const {
+    using detail::ParsedValueDesc;
+
+    const auto& values_buffer = values_buffers_.back();
+    const auto values = reinterpret_cast<const ParsedValueDesc*>(values_buffer->data());
+    const auto start_pos =
+        static_cast<int32_t>(values_buffer->size() / sizeof(ParsedValueDesc)) -
+        num_cols_ - 1;
+    for (int32_t col_index = 0; col_index < num_cols_; ++col_index) {
+      auto start = values[start_pos + col_index].offset;
+      auto stop = values[start_pos + col_index + 1].offset;
+      auto quoted = values[start_pos + col_index + 1].quoted;
+      ARROW_RETURN_NOT_OK(visit(parsed_ + start, stop - start, quoted));
+    }
+    return Status::OK();
+  }
+
+ protected:
+  // The number of rows in this batch
+  int32_t num_rows_ = 0;
+  // The number of columns
+  int32_t num_cols_ = 0;
+
+  // XXX should we ensure the parsed buffer is padded with 8 or 16 excess zero bytes?
+  // It may help with null parsing...
+  std::vector<std::shared_ptr<Buffer>> values_buffers_;
+  std::shared_ptr<Buffer> parsed_buffer_;
+  const uint8_t* parsed_ = NULLPTR;
+  int32_t parsed_size_ = 0;
+
+  friend class ::arrow::csv::BlockParserImpl;
+};
+
+}  // namespace detail
+
+constexpr int32_t kMaxParserNumRows = 100000;
 
 /// \class BlockParser
 /// \brief A reusable block-based parser for CSV data
@@ -62,6 +137,7 @@ class ARROW_EXPORT BlockParser {
                        int32_t max_num_rows = kMaxParserNumRows);
   explicit BlockParser(MemoryPool* pool, ParseOptions options, int32_t num_cols = -1,
                        int32_t max_num_rows = kMaxParserNumRows);
+  ~BlockParser();
 
   /// \brief Parse a block of data
   ///
@@ -86,11 +162,11 @@ class ARROW_EXPORT BlockParser {
   Status ParseFinal(const std::vector<util::string_view>& data, uint32_t* out_size);
 
   /// \brief Return the number of parsed rows
-  int32_t num_rows() const { return num_rows_; }
+  int32_t num_rows() const { return parsed_batch().num_rows(); }
   /// \brief Return the number of parsed columns
-  int32_t num_cols() const { return num_cols_; }
+  int32_t num_cols() const { return parsed_batch().num_cols(); }
   /// \brief Return the total size in bytes of parsed data
-  uint32_t num_bytes() const { return parsed_size_; }
+  uint32_t num_bytes() const { return parsed_batch().num_bytes(); }
 
   /// \brief Visit parsed values in a column
   ///
@@ -98,82 +174,18 @@ class ARROW_EXPORT BlockParser {
   /// Status(const uint8_t* data, uint32_t size, bool quoted)
   template <typename Visitor>
   Status VisitColumn(int32_t col_index, Visitor&& visit) const {
-    for (size_t buf_index = 0; buf_index < values_buffers_.size(); ++buf_index) {
-      const auto& values_buffer = values_buffers_[buf_index];
-      const auto values = reinterpret_cast<const ValueDesc*>(values_buffer->data());
-      const auto max_pos =
-          static_cast<int32_t>(values_buffer->size() / sizeof(ValueDesc)) - 1;
-      for (int32_t pos = col_index; pos < max_pos; pos += num_cols_) {
-        auto start = values[pos].offset;
-        auto stop = values[pos + 1].offset;
-        auto quoted = values[pos + 1].quoted;
-        ARROW_RETURN_NOT_OK(visit(parsed_ + start, stop - start, quoted));
-      }
-    }
-    return Status::OK();
+    return parsed_batch().VisitColumn(col_index, std::forward<Visitor>(visit));
   }
 
   template <typename Visitor>
   Status VisitLastRow(Visitor&& visit) const {
-    const auto& values_buffer = values_buffers_.back();
-    const auto values = reinterpret_cast<const ValueDesc*>(values_buffer->data());
-    const auto start_pos =
-        static_cast<int32_t>(values_buffer->size() / sizeof(ValueDesc)) - num_cols_ - 1;
-    for (int32_t col_index = 0; col_index < num_cols_; ++col_index) {
-      auto start = values[start_pos + col_index].offset;
-      auto stop = values[start_pos + col_index + 1].offset;
-      auto quoted = values[start_pos + col_index + 1].quoted;
-      ARROW_RETURN_NOT_OK(visit(parsed_ + start, stop - start, quoted));
-    }
-    return Status::OK();
+    return parsed_batch().VisitLastRow(std::forward<Visitor>(visit));
   }
 
  protected:
-  ARROW_DISALLOW_COPY_AND_ASSIGN(BlockParser);
+  std::unique_ptr<BlockParserImpl> impl_;
 
-  Status DoParse(const std::vector<util::string_view>& data, bool is_final,
-                 uint32_t* out_size);
-  template <typename SpecializedOptions>
-  Status DoParseSpecialized(const std::vector<util::string_view>& data, bool is_final,
-                            uint32_t* out_size);
-
-  template <typename SpecializedOptions, typename ValuesWriter, typename ParsedWriter>
-  Status ParseChunk(ValuesWriter* values_writer, ParsedWriter* parsed_writer,
-                    const char* data, const char* data_end, bool is_final,
-                    int32_t rows_in_chunk, const char** out_data, bool* finished_parsing);
-
-  // Parse a single line from the data pointer
-  template <typename SpecializedOptions, typename ValuesWriter, typename ParsedWriter>
-  Status ParseLine(ValuesWriter* values_writer, ParsedWriter* parsed_writer,
-                   const char* data, const char* data_end, bool is_final,
-                   const char** out_data);
-
-  MemoryPool* pool_;
-  const ParseOptions options_;
-  // The number of rows parsed from the block
-  int32_t num_rows_;
-  // The number of columns (can be -1 at start)
-  int32_t num_cols_;
-  // The maximum number of rows to parse from this block
-  int32_t max_num_rows_;
-
-  // Linear scratchpad for parsed values
-  struct ValueDesc {
-    uint32_t offset : 31;
-    bool quoted : 1;
-  };
-
-  // XXX should we ensure the parsed buffer is padded with 8 or 16 excess zero bytes?
-  // It may help with null parsing...
-  std::vector<std::shared_ptr<Buffer>> values_buffers_;
-  std::shared_ptr<Buffer> parsed_buffer_;
-  const uint8_t* parsed_;
-  int32_t values_size_;
-  int32_t parsed_size_;
-
-  class ResizableValuesWriter;
-  class PresizedValuesWriter;
-  class PresizedParsedWriter;
+  const detail::DataBatch& parsed_batch() const;
 };
 
 }  // namespace csv