… cpp/src/arrow/csv/parser.cc