@pitrou 2020-10-20
TODO: improve parser benchmarks a bit